requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
PySide6>=6.6
Pillow>=10.0
imagehash>=4.3
tqdm>=4.66
sentence-transformers
mediapipe
opencv-python
numpy
cupy-cuda12x>=13.0.0
torch==2.9.1+cu128
torchvision==0.24.1+cu128
ultralytics
pyexiv2
nvidia-cuda-runtime-cu12>=12.8.0
nvidia-cudnn-cu12>=9.7.0
nvidia-cublas-cu12>=12.8.0
